refactor: eliminate warning

+++ include/opengv/sac/MultiSampleConsensus.hpp
@@ -95,9 +95,9 @@ public:
   //private:
   
   /** the maximum number of iterations */
-  int max_iterations_;
+  size_t max_iterations_;
   /** the current number of iterations */
-  int current_iterations_;
+  size_t current_iterations_;
   /** the threshold for classifying inliers */
   double threshold_;
   /** the current probability (defines remaining iterations) */
